chore: rename module bv to bitvector
